requirements.txt
Flask==1.1.2
gunicorn==20.0.4
jinja2==2.11.2
itsdangerous==1.1.0
MarkupSafe==1.1.1
Werkzeug==1.0.1
numpy==1.16.4
pandas==1.0.3
pytesseract==0.3.4
opencv==3.4.1
opencv-python==4.2.0.34
sqlite==3.31.1
pillow==7.1.2
tesseract==0.1.3